Added casting warning message

+++ Runtime/Math/Math.h
@@ -5,6 +5,6 @@ namespace Mile
 {
    namespace Math
    {
-      static const Float Pi = 3.1415926535897932384;
+      static const Float Pi = (Float)3.1415926535897932384;
    }
 }
